see prior commit, forgot to update header

+++ src/Panel_HUB75E.h
@@ -282,7 +282,7 @@ public:
     void swapBuffer(const LED *newBuffer, uint8_t bufferLength);
     void fillBuffer(uint16_t color);
     void drawRect(uint8_t x1, uint8_t y1, uint8_t x2, uint8_t y2, uint16_t color, bool fill);
-    void drawSquare(uint8_t x, uint8_t y, uint8_t size, uint8_t color, bool fill);
+    void drawSquare(uint8_t x, uint8_t y, uint8_t size, uint16_t color, bool fill);
     void drawLine(uint8_t x1, uint8_t y1, uint8_t x2, uint8_t y2, uint16_t color);
     void drawEllipse(uint8_t xm, uint8_t ym, uint8_t a, uint8_t b, uint16_t color, bool fill);
     void drawCircle(uint8_t xm, uint8_t ym, uint8_t radius, uint16_t color, bool fill);
